// frontend/src/app/core/interfaces/models/post.model.interface.ts
import { ICategory } from "./category.model.interface";
import { IUser } from "./user.model.interface";

export interface IPost {
    id: number;
    title: string;
    content: string;
    slug: string;
    userId: number;  //?
    user: IUser;
    categoryId: number;
    category: ICategory;
    createdAt: string;
    updatedAt: string;
    totalComments?: number;
}